Fixed incorrect row/height issue

+++ Player.py
@@ -78,18 +78,20 @@ class Player(arcade.Sprite):
     def dig_down(self):
         block_to_dig = self.map.get_block(self.row + 1, self.col)
 
-        if block_to_dig.block_type is "stone":
-            self.map.destroy_map_block(self.row + 1, self.col)
-            self.row += 1
-            self.update_player_position()
-            self.check_if_player_should_fall()
-       
-        elif block_to_dig.block_type is "stone_gold":
-            self.map.destroy_map_block(self.row + 1, self.col)
-            self.score += 5
-            self.row += 1
-            self.update_player_position()
-            self.check_if_player_should_fall()
+        if block_to_dig is not None:
+
+            if block_to_dig.block_type is "stone":
+                self.map.destroy_map_block(self.row + 1, self.col)
+                self.row += 1
+                self.update_player_position()
+                self.check_if_player_should_fall()
+
+            elif block_to_dig.block_type is "stone_gold":
+                self.map.destroy_map_block(self.row + 1, self.col)
+                self.score += 5
+                self.row += 1
+                self.update_player_position()
+                self.check_if_player_should_fall()
 
     def check_if_player_should_fall(self):
         block_to_check = self.map.get_block(self.row + 1, self.col)
